refactor: 로그인 사용자 정보 조회 메서드를 AuthController에서 UserController로 이동
